+++ server/actions/handle-batch.js
@@ -3,7 +3,7 @@ import Promise from "bluebird";
 
 export default function handleBatch(ctx, messages) {
   const users = messages.map(m => {
-    const segmentIds = m.segments.map(s => s.id);
+    const segmentIds = _.compact(m.segments).map(s => s.id);
     m.user.segment_ids = _.compact(_.uniq((m.user.segment_ids || []).concat(segmentIds)));
     return m.user;
   });