+++ server/controller/notify.js
@@ -5,7 +5,7 @@ export default class UserUpdateStrategy {
   static userUpdateHandler(ctx, messages) {
     const { syncAgent } = ctx.shipApp;
     if (!syncAgent.isConfigured()) {
-      ctx.client.logger.info("ship is not configured");
+      ctx.client.logger.error("connector.configuration.error", { errors: "connector is not configured" });
       return Promise.resolve();
     }
 
@@ -33,7 +33,7 @@ export default class UserUpdateStrategy {
   static shipUpdateHandler(ctx) {
     const { syncAgent } = ctx.shipApp;
     if (!syncAgent.isConfigured()) {
-      ctx.client.logger.info("ship is not configured");
+      ctx.client.logger.error("connector.configuration.error", { errors: "connector is not configured" });
       return Promise.resolve();
     }
     return ctx.enqueue("shipUpdateJob");
@@ -42,20 +42,24 @@ export default class UserUpdateStrategy {
   static segmentUpdateHandler(ctx, message) {
     const { syncAgent } = ctx.shipApp;
     if (!syncAgent.isConfigured()) {
-      ctx.client.logger.info("ship is not configured");
+      ctx.client.logger.error("connector.configuration.error", { errors: "connector is not configured" });
       return Promise.resolve();
     }
     const segment = message;
     return syncAgent.setupShip()
       .then(() => {
         return ctx.helpers.requestExtract({ segment, fields: syncAgent.mapping.getHullTraitsKeys() });
+      })
+      .catch((err) => {
+        ctx.client.logger.error("requestExtract.error", err);
+        return Promise.resolve("err");
       });
   }
 
   static segmentDeleteHandler(ctx) {
     const { syncAgent } = ctx.shipApp;
     if (!syncAgent.isConfigured()) {
-      ctx.client.logger.info("ship is not configured");
+      ctx.client.logger.error("connector.configuration.error", { errors: "connector is not configured" });
       return Promise.resolve();
     }
     // TODO: if the segment would have `query` param we could trigger an extract
@@ -69,6 +73,10 @@ export default class UserUpdateStrategy {
         return Promise.map(segments, (segmentId) => {
           return ctx.helpers.requestExtract({ segment: { id: segmentId }, remove: true, fields: syncAgent.mapping.getHullTraitsKeys() });
         });
+      })
+      .catch((err) => {
+        ctx.client.logger.error("requestExtract.error", err);
+        return Promise.resolve("err");
       });
   }
 }