+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## v0.3.7
+- fix segment parsing
+- upgrade hull-node to 0.11.8 -> new firehose
+- handle requestExtract errors
+- update logging convention
+
 ## v0.3.6
 - upgrade hull-node to 0.11.3
 - fix batch handler